Revert "v0.63 sync from upstream (#613)"
This reverts commit efc221bd10cd8bf8fcf675eaec6a47f9bf5d385c.

+++ ReactCommon/fabric/components/scrollview/ScrollViewProps.h
@@ -44,7 +44,7 @@ class ScrollViewProps final : public ViewProps {
   const Float zoomScale{1.0};
   const EdgeInsets contentInset{};
   const EdgeInsets scrollIndicatorInsets{};
-  const Float snapToInterval{};
+  const int snapToInterval{};
   const ScrollViewSnapToAlignment snapToAlignment{};
 
 #pragma mark - DebugStringConvertible
